pyproject.toml
[project]
name = "rag-demo-unified"
version = "0.1.0"
description = "Unified multimodal PDF RAG demo with Docling, Qdrant, OpenAI, and Gradio — combines the best of rag-demo-claude, rag-demo-codex, and rag-demo-gemini"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "gradio>=5.0.0",
    "qdrant-client>=1.11.0",
    "openai>=1.50.0",
    "pillow>=10.0.0",
    "pymupdf>=1.24.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
    "pydantic>=2.0.0",
    "requests>=2.32.0",
]

[project.optional-dependencies]
docling = ["docling>=2.0.0", "docling-core>=2.0.0"]

[project.scripts]
rag-demo = "rag_demo.cli:main"

[dependency-groups]
dev = ["pytest>=8.0.0"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_demo"]
